% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AovTukeySummary.R
\name{Aov1TukeySummary}
\alias{Aov1TukeySummary}
\title{Aov1TukeySummary}
\usage{
Aov1TukeySummary(df, Vari., Factor, DeciNum = rep(2, length(Vari.)))
}
\arguments{
\item{df}{A data.frame}

\item{Vari.}{A character vacter, column names of the variables}

\item{Factor}{The column name of the factor}

\item{DeciNum}{A numeric vector of decimal numbers, default is 2}
}
\value{
A data.frame
}
\description{
Summarize one-way anova and Tukey test results
Gives the mean value, grouping label, probability including interaction, MSD value
}
\examples{
Omitted
}
