% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AovTukeySummary.R
\name{Aov2TukeySummary}
\alias{Aov2TukeySummary}
\title{Summarize two-way anova and Tukey test results
Gives the mean value, grouping label, probability including interactions, MSD value, Eta_square}
\usage{
Aov2TukeySummary(df, Vari., Factor1, Factor2, DeciNum = rep(2, length(Vari.)))
}
\arguments{
\item{df}{A data.frame}

\item{Vari.}{A character vacter, column names of the variables}

\item{Factor1}{The column name of the factor1}

\item{Factor2}{The column name of the factor2}

\item{DeciNum}{A numeric vector of decimal numbers, default is 2}
}
\value{
A data.frame
}
\description{
Summarize two-way anova and Tukey test results
Gives the mean value, grouping label, probability including interactions, MSD value, Eta_square
}
\examples{
Omitted
}
