% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bruker2conc.R
\name{Bruker2conc}
\alias{Bruker2conc}
\title{Bruker2conc}
\usage{
Bruker2conc(
  BrukerDir,
  InRef = "TSP",
  InRef2 = "4-Methylvalerate",
  CorInRef2 = T,
  ConcRef2 = 4,
  CorChemshift2Zero = T,
  Samp2run = "All",
  ncolplot = 5,
  EndProductInfo = "EndProductInfo.xlsx",
  OutDir = getwd()
)
}
\arguments{
\item{BrukerDir}{A Character. Path to the folder where all the Bruker out directories are stored}

\item{InRef}{A character. Internal standard. Default - "TSP"}

\item{InRef2}{A character OR False. A second internal standard. Default = "4-Methylvalerate"}

\item{CorInRef2}{Logical. If the concentration should be corrected by the second internal standard. Default - T.}

\item{ConcRef2}{Numeric. Concentration of the second internal standard in nM. Only works if CorInRef2 = T. Default - 4.}

\item{CorChemshift2Zero}{Logical. If the chemical shift should be corrected. If T, all chemical shift will be shifting so that for InRef it is 0. Default - T.}

\item{Samp2run}{"All" or a character vector containing all samples names to process. Default - "All"}

\item{ncolplot}{Numeric. number of column in the intergral detail plots. Default - 5.}

\item{EndProductInfo}{Character. Path to the EndProductInfo.xlsx file. This file should include two sheets, namely ChemRange and Calculation. Default - "EndProductInfo.xlsx"}

\item{OutDir}{Character. Path to export resulting files. Default - getwd}
}
\description{
Convert Bruker output files to final concentration. All spectra should be at least displayed once in Topspin.
An EndproductInfo.xlsx file is requried to include the chemical shift range and the calculation information. Please download a demo from
my github repository hyfood/DeomoData/EndproductInfo.xlsx. The DemoData Directory also includes a Bruker output directory as a demo for this function.
}
