% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClearTaxon.R
\name{ClearGenusSilva138}
\alias{ClearGenusSilva138}
\title{ClearGenusSilva138}
\usage{
ClearGenusSilva138(raw_taxon)
}
\arguments{
\item{raw_taxon}{A character vector of taxa, in form of k__xxxx;p__xxxx;c__xxxx;o__xxxx;f__xxxx;g__xxxx}
}
\value{
Returns a character vector of the genus taxa
}
\description{
Make clear and short genus name(s) for taxon based on SILVA 138 (i.e., k__;p__;c__;o__;f__;g__)
}
\examples{
raw_taxon <- c("Unassigned;__;__;__;__;__",
"k__Archaea;p__Euryarchaeota;c__Methanobacteria;o__Methanobacteriales;f__Methanobacteriaceae;g__Methanobrevibacter",
"k__Bacteria;p__Actinobacteriota;c__Coriobacteriia;o__Coriobacteriales;f__Coriobacteriales_Incertae_Sedis;g__uncultured")
new_taxon <- ClearGenusSilva138(raw_taxon)
}
