% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doBiplot.R
\name{doBiplot}
\alias{doBiplot}
\title{doBiplot}
\usage{
doBiplot(
  toBiplotList,
  GrpInfo = NULL,
  AxisVector = c(1, 2),
  Vari2Arrow = NULL,
  AxisTitle = NA,
  Decimal = 2,
  GrpCol = NULL,
  GrpShape = NULL,
  GrpSize = NULL,
  NoGrpCol = "black",
  NoGrpShape = "circle",
  NoGrpSize = 3,
  PointAlpha = 0.5,
  PointText = NULL,
  Colmannual = NULL,
  Shapemannual = NULL,
  Sizemannual = NULL,
  DrawElips = F,
  GrpElips = NULL,
  ElipsCol = NULL,
  ShowArrow = T,
  ArrowSize = 1,
  ArrowLength = 0.3,
  SegLengthMultiply = NULL,
  ArrowAlpha = 1,
  ArrowCol = "black",
  ArrowTextSize = 3,
  axis.title = 15,
  axis.text.x = 15,
  axis.text.y = 15,
  legend.key.size = 1,
  legend.title = 16,
  legend.text = 16
)
}
\arguments{
\item{toBiplotList}{The output list of PCAtoBiplot or PCoAtoBiplot.}

\item{GrpInfo}{Grouping Info data.frame if the color, size, shape and elipse need to be defined. Same rownames(sample names) as the toBiplotList[\link{1}]. Default - NULL.}

\item{AxisVector}{Numeric vector with length of 2, wo define which two axis to do the biplot. Default - c(1,2).}

\item{Vari2Arrow}{Character vector to determine which variable(s) to show on the biplot as arrow(s).}

\item{AxisTitle}{A Character. Name of Axis title. e.g. "PCA" or "PCoA" or "Axis".}

\item{Decimal}{Numeric. Decimal points of axis title percentage. Default - 2.}

\item{GrpCol}{Character. Which group (column) is used to define the color. Default - NULL.}

\item{GrpShape}{Character. Which group (column) is used to define the shape. Default - NULL.}

\item{GrpSize}{Character. Which group (column) is used to define the size. Default - NULL.}

\item{NoGrpCol}{Character. Color to use if not define the GrpCol. Default - "black".}

\item{NoGrpShape}{Character. Shape to use if not define the GrpCol. Default - "circle".}

\item{NoGrpSize}{Numeric. Size to use if not define the GrpCol. Default - 3.}

\item{PointAlpha}{Numeric. Transparency of points. Default - 0.5.}

\item{PointText}{Character. Which column is used to show the text on plot. Default - NULL.}

\item{Colmannual}{Character vector. Define the color. Default - NULL.}

\item{Shapemannual}{Character vector. Define the shape. Default - NULL.}

\item{Sizemannual}{Character vector. Define the size. Default - NULL.}

\item{DrawElips}{Logical. Draw elipse or not.Default - T.}

\item{GrpElips}{Character. Which group (column) is used to Draw elipse. Default - NULL.}

\item{ElipsCol}{Character. Which group (column) is used to define elipse color. Default - NULL.}

\item{ShowArrow}{Logical. Draw arrows or not. Default - T}

\item{ArrowSize}{Numeric. Arrow size. Default - 1.}

\item{ArrowLength}{Numeric. Arrow length. Default - 0.3.}

\item{SegLengthMultiply}{Numeric. Length factor to multiply to the segment. Default - NULL.}

\item{ArrowAlpha}{Numeric. Arrow transparency. Default = 1.}

\item{ArrowCol}{Character. Arrow color. Default - "black".}

\item{ArrowTextSize}{Numeric. Text size of arrow. Default - 3.}

\item{axis.title}{Numeric. Text size of axis title. Default - 15.}

\item{axis.text.x}{Numeric. Text size of x axis. Default - 15.}

\item{axis.text.y}{Numeric. Text size of y axis. Default - 15.}

\item{legend.key.size}{Numeric. Legend key size. Default - 1.}

\item{legend.title}{Numeric. Legend title size. Default - 16.}

\item{legend.text}{Numeric. Legend text size Default - 16.}
}
\value{
This function returns a ggplot object for biplot.
}
\description{
This function returns a ggplot object for biplot.
}
\examples{
Coming...
}
