% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DonutAov.R
\name{DonutAov}
\alias{DonutAov}
\title{DonutAov}
\usage{
DonutAov(
  aovdf,
  Vari.,
  single_out = F,
  all_out = F,
  all_out_col = 1,
  outfolder = getwd(),
  res = 150,
  height = 1000,
  width = 1000,
  VariSize = 8,
  labelSize = 4,
  DonutFill = "default"
)
}
\arguments{
\item{aovdf}{The data.frame output from Aov2TukeySummary OR Aov3TukeySummary function}

\item{Vari.}{A character vector contaning variables to make donut charts}

\item{single_out}{Logical. Whether to export every single Donut .png file. Default - F}

\item{all_out}{Logical. Whether to export all Donut figures in a .png file. Default - F}

\item{all_out_col}{Numeric. Column numbers in the all_out .png file. Default - 1}

\item{outfolder}{Character. Folder path to export the .png file. default - getwd. Better to setwd prior to this function}

\item{res}{Numeric. Resolution of the output .png file. Default - 150}

\item{height}{Numeric. Height of each single .png file. Default - 1000}

\item{width}{Numeric. Width of each single .png file. Default - 1000}

\item{VariSize}{Numeric. Font size of the Vari. on Donut chart. Default - 8}

\item{labelSize}{Numeric. Font size of the labels, i.e. Variable&Interaction&Residuals. Default - 4}

\item{DonutFill}{Character vector must have a length of 7 OR 10. Colors to choose for donut chart. Anti-clockwise, outer donut and then inner donut. Default - Default.}
}
\value{
One or more .png files
}
\description{
Making Donut charts summarising the Eta2 of two-way or three-way ANOVA based on Aov2TukeySummary OR Aov2TukeySummary output data.frame
}
\examples{
Omitted
}
