% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterTop.R
\name{FilterTop}
\alias{FilterTop}
\title{FilterTop}
\usage{
FilterTop(df, TaxonCol = NULL, AbunCol, Others = F, Threshold)
}
\arguments{
\item{df}{A data.frame. Variables (e.g. taxa) as rows; Samples as colnames.}

\item{TaxonCol}{A character. The colname of the Taxon column. Should be specified if Others = T.}

\item{AbunCol}{A numeric/character vector, or a number/character, of the collumn number(s)/name(s), to transfer to relative abundance}

\item{Others}{Logical. If T, an "Others" row will be generated.}

\item{Threshold}{A number. The threshold that the min values across all collumns less than it will not be included in the output data.frame.}
}
\value{
Returns a data.frame with the top abundant taxa
}
\description{
This functions filters the top abundant data with a specified threshold
}
\examples{
library(FermentKinetics)
df <- data.frame(Taxon = paste0("Taxon_",1:10),
                x = abs(rnorm(10,2)),
                y = abs(rnorm(10,1.5)),
                z = abs(rnorm(10,3)))
df <- RelAbun(df, AbunCol = c(2:4))
Top5df <- FilterTop(df, TaxonCol = "Taxon",AbunCol = c(2:4), Threshold = 5)

}
