% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitFranceModel.R
\name{FitFranceModel}
\alias{FitFranceModel}
\title{FitGrootModel}
\usage{
FitFranceModel(x, y)
}
\arguments{
\item{x}{A numeric vector. Time}

\item{y}{A numeric vector. gas volume}
}
\value{
Returns a vactor with all parameters. A: asymptotiv value. Rmax: maximum gas production rate.
TRmax: time to reach RMax. LagT: lag phase time. FinalCV: final cumulative volume
}
\description{
This function fits gas data to France monophasic Model: A*(1 - exp(-b*(time - T) - c*(sqrt(time)-sqrt(T))))
Dhanoa, M. S., López, S., Powell, C. D., Sanderson, R., Ellis, J. L., Murray, J.-A., Garber, A., Williams, B. A., & France, J. (2021). An illustrative analysis of atypical gas production profiles obtained from in vitro digestibility studies using fecal inoculum. Animals, 11(4), 1069.
}
\examples{
Coming.
}
