% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGrootModel.R
\name{FitGrootModel}
\alias{FitGrootModel}
\title{FitGrootModel}
\usage{
FitGrootModel(x, y)
}
\arguments{
\item{x}{A numeric vector. Time}

\item{y}{A numeric vector. gas volume}
}
\value{
Returns a vactor with all parameters. A: asymptotiv value. C: half time to reach A. B: switching characteristics. Rmax: maximum gas production rate.
TRmax: time to reach RMax. LagT: lag phase time. FinalCV: final cumulative volume
}
\description{
This function fits gas data to Groot Model: A/(1+(C/time)^B).
Please read:Groot, J. C., Cone, J. W., Williams, B. A., Debersaques, F. M., & Lantinga, E. A. (1996). Multiphasic analysis of gas production kinetics for in vitro fermentation of ruminant feeds. Animal Feed Science Technology, 64(1), 77-89.
I also added a LagT defined by the intercept of y = 0 and the tangent line passing RMax.
}
\examples{
Coming.
}
