% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gapfill.R
\name{Gapfill}
\alias{Gapfill}
\title{Gapfill}
\usage{
Gapfill(Time, gasV, Total_t, gap)
}
\arguments{
\item{Time}{Numeric vector. Time points}

\item{gasV}{Numeric vector. Gas volume data.}

\item{Total_t}{A number. Total time, can be greater than the max(Time).}

\item{gap}{A number. The minimum time gap.}
}
\value{
Returns a list. List[\link{1}] is the filled time points. List[\link{2}] is the filled gas volume.
}
\description{
This function fills the gaps for the time-gas data by applying linear functions. This step should be done prior to curve-fitting, to balance to weights of the measured points.
}
\examples{
Time <- c(0,2.731,3.618,4.101,4.426,4.695,
                    4.996,5.286,5.602,5.933,6.274,6.612,
                    6.996,7.451,8.136,10.552,46.771)
gasV <- c(0,2.13,  4.16,  6.24,  8.40, 10.55, 12.69, 14.94, 17.25, 19.41, 21.54, 23.64,
          25.71, 27.80, 29.85, 31.85, 33.85)
Filled <- Gapfill(Time, gasV, 48, 5/60)
plot(Filled[[1]], Filled[[2]])
}
