% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAbiplot.R
\name{PCAbiplot}
\alias{PCAbiplot}
\title{PCAbiplot}
\usage{
PCAbiplot(
  df,
  Vari.,
  scale = T,
  center = T,
  Grps = NULL,
  GrpCol = NULL,
  GrpColLevels = NULL,
  GrpShape = NULL,
  GrpShapeLevels = NULL,
  GrpSize = NULL,
  GrpSizeLevels = NULL,
  NoGrpCol = "black",
  NoGrpShape = "circle",
  NoGrpSize = 3,
  PointAlpha = 0.5,
  PointText = NULL,
  Colmannual = NULL,
  Shapemannual = NULL,
  Sizemannual = NULL,
  DrawElips = F,
  GrpElips = NULL,
  ElipsCol = NULL,
  ShowArrow = T,
  ArrowSize = 1,
  ArrowLength = 0.3,
  SegLengthMultiply = NULL,
  ArrowAlpha = 1,
  ArrowCol = "black",
  ArrowTextSize = 3,
  axis.title = 15,
  axis.text.x = 15,
  axis.text.y = 15,
  legend.key.size = 1,
  legend.title = 16,
  legend.text = 16
)
}
\arguments{
\item{df}{A data.frame. Can include the grouping info column(s).}

\item{Vari.}{Character vector. Which variables (columns) are used.}

\item{scale}{Logical. Default - T.}

\item{center}{Logical. Default - T.}

\item{Grps}{Character vector. Which groups (columns) are used. Default - NULL.}

\item{GrpCol}{Character. Which group (column) is used to define the color. Default - NULL.}

\item{GrpColLevels}{Character vector, setting the levels of the GrpCol. Default - NULL.}

\item{GrpShape}{Character. Which group (column) is used to define the shape. Default - NULL.}

\item{GrpShapeLevels}{Character vector, setting the levels of the GrpShape. Default - NULL.}

\item{GrpSize}{Character. Which group (column) is used to define the size. Default - NULL.}

\item{GrpSizeLevels}{Character vector, setting the levels of the GrpSize. Default - NULL.}

\item{NoGrpCol}{Character. Color to use if not define the GrpCol. Default - "black".}

\item{NoGrpShape}{Character. Shape to use if not define the GrpCol. Default - "circle".}

\item{NoGrpSize}{Numeric. Size to use if not define the GrpCol. Default - 3.}

\item{PointAlpha}{Numeric. Transparency of points. Default - 0.5.}

\item{PointText}{Character. Which column is used to show the text on plot. Default - NULL.}

\item{Colmannual}{Character vector. Define the color. Default - NULL.}

\item{Shapemannual}{Character vector. Define the shape. Default - NULL.}

\item{Sizemannual}{Character vector. Define the size. Default - NULL.}

\item{DrawElips}{Logical. Draw elipse or not.Default - T.}

\item{GrpElips}{Character. Which group (column) is used to Draw elipse. Default - NULL.}

\item{ElipsCol}{Character. Which group (column) is used to define elipse color. Default - NULL.}

\item{ShowArrow}{Logical. Draw arrows or not. Default - T}

\item{ArrowSize}{Numeric. Arrow size. Default - 1.}

\item{ArrowLength}{Numeric. Arrow length. Default - 0.3.}

\item{SegLengthMultiply}{Numeric. Length factor to multiply to the segment. Default - NULL.}

\item{ArrowAlpha}{Numeric. Arrow transparency. Default = 1.}

\item{ArrowCol}{Character. Arrow color. Default - "black".}

\item{ArrowTextSize}{Numeric. Text size of arrow. Default - 3.}

\item{axis.title}{Numeric. Text size of axis title. Default - 15.}

\item{axis.text.x}{Numeric. Text size of x axis. Default - 15.}

\item{axis.text.y}{Numeric. Text size of y axis. Default - 15.}

\item{legend.key.size}{Numeric. Legend key size. Default - 1.}

\item{legend.title}{Numeric. Legend title size. Default - 16.}

\item{legend.text}{Numeric. Legend text size Default - 16.}
}
\value{
Returns a list. See description
}
\description{
This function returns a list for information of PCA biplot, and a ggplot object (list[\link{4}] in the list) which can be export directly. Otherwise, you can generate your own
ggplot object just from the first three elements. list[\link{1}] is a data.frame showing the point for PCA plot and the grouping information.
list[\link{2}] is a vector of the percentages of each PC. list[\link{3}] is a data.frame for loadings (rotation). list[\link{4}] is a data.frame giving the information from which you can generate the arrows by: geom_segment(data=list[\link{3}], aes(x=0, y=0, xend=v1, yend=v2))
}
\examples{
Download a demo data from my github: hyfood/DemoData/PCAbiplotDemo.txt, and read into R.
PCAbiplotDemo$Dnr <- as.character(PCAbiplotDemo$Dnr)
biplotList <- PCAbiplot(df = PCAbiplotDemo,
Vari. = c("A", "B", "C", "cCV", "TRMax", "RMax", "LagT"),
Grps = c("Substrate", "Dnr", "PWgt"),
GrpCol = "Dnr",
GrpShape = "Substrate",
GrpSize = "PWgt",
DrawElips = T,
GrpElips = "Dnr",
ElipsCol = "Dnr")
}
