% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAtoBiplot.R
\name{PCAtoBiplot}
\alias{PCAtoBiplot}
\title{PCAtoBiplot}
\usage{
PCAtoBiplot(df, scale = T, center = T)
}
\arguments{
\item{df}{Numeric data.frame. Sample names as rownames and variables as colnames.}

\item{scale}{Logical. Default - T.}

\item{center}{Logical. Default - T.}
}
\value{
Returns a list. See description
}
\description{
This function returns a list containing elements required for PCA biplot. list[\link{1}] is the PCA axis data. list[\link{2}] is a vector including the PC variation explained percentages.
list [\link{3}] is the loadings
}
\examples{
None.
}
