% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelAbun.R
\name{RelAbun}
\alias{RelAbun}
\title{RelAbun}
\usage{
RelAbun(df, AbunCol)
}
\arguments{
\item{df}{a data.frame, with sample name(s) as colname(s)}

\item{AbunCol}{A numeric/character vector, or a number/character, of the collumn number(s)/name(s), to transfer to relative abundance}
}
\value{
Returns a relative abundace data.frame
}
\description{
Convert an input data.frame to relative abundance \%
}
\examples{
df <- data.frame(Taxon = paste0("Taxon_",1:30),
                x = abs(rnorm(30)),
                y = abs(rnorm(30)),
                z = abs(rnorm(30)))
                Reldf <- RelAbun(df, c(2:4))


}
